corrigiendo fallas de seguridad del proyecto, terminando funcionalidad

+++ todo/todo.py
@@ -12,7 +12,11 @@ bp = Blueprint('todo', __name__)
 def index():
     db, c = get_db()
     c.execute(
-        "select t.id, t.description, u.username, t.completed, t.created_at from todo t JOIN user u on t.created_by = u.id where u.id order by created_at desc"
+        """
+        select t.id, t.description, t.completed, t.created_by, t.created_at, u.username
+        from todo t JOIN user u on t.created_by = u.id where t.created_by = %s order by t.created_at desc
+        """,
+        (g.user['id'],)
     )
     todos = c.fetchall()
     return render_template('todo/index.html', todos=todos)
@@ -56,8 +60,8 @@ def update(id):
         else:
             db, c = get_db()
             c.execute(
-                'UPDATE todo SET description = %s, completed = %s WHERE id = %s',
-                (description, completed, id)
+                'UPDATE todo SET description = %s, completed = %s WHERE id = %s AND created_by = %s',
+                (description, completed, id, g.user['id'])
             )
             db.commit()
             return redirect(url_for('todo.index'))
@@ -67,7 +71,7 @@ def update(id):
 @login_required
 def delete(id):
     db, c = get_db()
-    c.execute('DELETE FROM todo WHERE id = %s', (id,))
+    c.execute('DELETE FROM todo WHERE id = %s and created_by = %s', (id, g.user['id']))
     db.commit()
     return redirect(url_for('todo.index'))
